file

+++ controllers/HostController.js
@@ -10,13 +10,10 @@ exports.getAddHome = (req, res) => {
   });
 };
 
-
 // GET Host Hostels List
 exports.getHostHostels = (req, res) => {
-
   Home.find()
     .then((registeredHomes) => {
-
       res.render("host/hostels", {
         registeredHomes,
         PageTitle: "Host Homes List",
@@ -24,16 +21,17 @@ exports.getHostHostels = (req, res) => {
         isLoggedIn: req.isLoggedIn,
         user: req.session.user
       });
-
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      console.error("Error fetching homes:", err);
+      res.status(500).send("Something went wrong!");
+    });
 };
-
 
 // POST Add Home
 exports.postAddHome = (req, res) => {
-
-  const { homeName, address, city, price, image } = req.body;
+  const { homeName, address, city, price } = req.body;
+  const image = req.file ? req.file.filename : null; // multer file check
 
   const home = new Home({
     homeName,
@@ -48,35 +46,45 @@ exports.postAddHome = (req, res) => {
       console.log("Home Saved Successfully");
       res.redirect("/submit-home");
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      console.error("Error saving home:", err);
+      res.status(500).send("Something went wrong!");
+    });
 };
-
 
 // SUCCESS PAGE
 exports.getSuccess = (req, res) => {
-
   res.render("submit-home", {
     PageTitle: "Home Submitted",
     cssFile: "submit-home",
     isLoggedIn: req.isLoggedIn,
     user: req.session.user
   });
-
 };
-
 
 // DELETE HOSTEL
 exports.deleteHostel = (req, res) => {
-
   const hostelId = req.body.hostelId;
 
+  if (!hostelId) {
+    console.error("No hostelId provided for deletion");
+    return res.status(400).send("Bad Request");
+  }
+
   Home.findByIdAndDelete(hostelId)
-    .then(() => {
+    .then((deletedHome) => {
+      if (!deletedHome) {
+        console.warn("Hostel not found for deletion:", hostelId);
+        return res.status(404).send("Hostel not found");
+      }
 
-      console.log("Hostel Deleted");
+      console.log("Hostel Deleted:", deletedHome._id);
 
-      res.redirect("/host/hostels");
-
+      //  Redirect to the correct route
+      res.redirect("hostels");
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+      console.error("Error deleting hostel:", err);
+      res.status(500).send("Something went wrong!");
+    });
 };
